[refactor] reorganiza rotas do backend
