+++ backend/src/controllers/transaction.controller.ts
@@ -4,6 +4,7 @@ import UserService from "../services/user.service";
 import AccountService from "../services/account.service";
 import statusCodes from "../helpers/statusCodes";
 import sequelize from '../database/models';
+import { IUser } from '../interfaces/IUser';
 
 export default class TransactionController {
   constructor(
@@ -41,5 +42,55 @@ export default class TransactionController {
       .status(statusCodes.INTERNAL_ERROR)
       .json({ message: error.message });
     }
+  }
+
+  public async getTransactionsByUsername(req: Request, res: Response) {
+    const { username } = req.params;
+
+    const user = await this._userService.getOneByUsername(username);
+
+    if (!user) {
+      return res.status(statusCodes.NOT_FOUND).json({
+        message: 'User not found',
+      });
+    }
+
+    const transactions = await this.getAllTransactions(user);
+
+    return res.status(statusCodes.OK).json(transactions);
+  }
+
+  public async filterTransactionsByDate(req: Request, res: Response) {
+    const { username } = req.params;
+    const { date } = req.body;
+
+    const user = await this._userService.getOneByUsername(username);
+
+    if (!user) {
+      return res.status(statusCodes.NOT_FOUND).json({
+        message: 'User not found',
+      });
+    }
+
+    const transactions = await this
+      .getTransactionsByDate(user, date);
+
+    return res.status(statusCodes.OK).json(transactions);
+  }
+
+  private async getAllTransactions(user: IUser) {
+    const cashIn = await this._transactionService
+      .getCashIn(user.accountId as number);
+    const cashOut = await this._transactionService
+      .getCashOut(user.accountId as number);
+    return { cashIn, cashOut };
+  }
+
+  private async getTransactionsByDate(user: IUser, date: string) {
+    const cashIn = await this._transactionService
+      .getCashInByDate(user.accountId as number, date);
+    const cashOut = await this._transactionService
+      .getCashOutByDate(user.accountId as number, date);
+    return { cashIn, cashOut };
   }
 }
