+++ backend/src/routes/user.ts
@@ -11,16 +11,6 @@ router
     '/user/:username',
     auth,
     (req: Request, res: Response) => controller.getByUsername(req, res),
-  )
-  .get(
-    '/user/:username/transactions',
-    auth,
-    (req: Request, res: Response) => controller.getTransactionsByUsername(req, res),
-  )
-  .get(
-    '/user/:username/transactions/filter-by-date',
-    auth,
-    (req: Request, res: Response) => controller.filterTransactionsByDate(req, res),
   );
 
 export default router;
